Refacotr _main as private method #main
A parsing error is thrown but it executes correcly
